+++ src/pyraingen/dailysequences.py
@@ -78,7 +78,7 @@ def dailySequences(nSeasons,
             # Grab a conveniance variable:
             currStnIndex = int(nearStationIdx[loopSeason, loopStation])
 
-            if currStnIndex == 0:
+            if currStnIndex < 0:
             # There are no more stations for this season
                 break
             else:
@@ -96,6 +96,7 @@ def dailySequences(nSeasons,
                 yearStart = int(dayVecStart[0])
                 yearEnd = int(dayVecEnd[0])
                 tmpSubDaily = ds['rainfall'][:].data/10
+                ds.close()
                 # The algorithm below works on the assumption that the
                 # tmpSubDaily array is populated with full years.  So pad out
                 # the data array to make full years with missingDay values.